ReplayGain FTW!
Make replay gain support actually do something by
(a) getting the data we stored out of the collection database
(b) using it when the track changes

Also, improve the storage of replay gain tags by storing NULL when they weren't present on the original track metadata.  This allows us to substitute the track gain for the album gain when the latter is requested but doesn't exist.


svn path=/trunk/extragear/multimedia/amarok/; revision=911684

+++ src/collection/sqlcollection/ScanManager.cpp
@@ -560,17 +560,17 @@ XmlParseJob::run()
                         data.insert( Meta::Field::FILESIZE, attrs.value( "filesize" ).toString() );
                     if( !attrs.value( "albumgain" ).isEmpty() )
                     {
-                        data.insert( Meta::Field::ALBUMGAIN, attrs.value( "albumgain" ).toString() );
+                        data.insert( Meta::Field::ALBUMGAIN, attrs.value( "albumgain" ).toString().toDouble() );
                         // peak gain is in decibels
                         if( !attrs.value( "albumpeakgain" ).isEmpty() )
-                            data.insert( Meta::Field::ALBUMPEAKGAIN, attrs.value( "albumpeakgain" ).toString() );
+                            data.insert( Meta::Field::ALBUMPEAKGAIN, attrs.value( "albumpeakgain" ).toString().toDouble() );
                     }
                     if( !attrs.value( "trackgain" ).isEmpty() )
                     {
-                        data.insert( Meta::Field::TRACKGAIN, attrs.value( "trackgain" ).toString() );
+                        data.insert( Meta::Field::TRACKGAIN, attrs.value( "trackgain" ).toString().toDouble() );
                         // peak gain is in decibels
                         if( !attrs.value( "trackpeakgain" ).isEmpty() )
-                            data.insert( Meta::Field::TRACKPEAKGAIN, attrs.value( "trackpeakgain" ).toString() );
+                            data.insert( Meta::Field::TRACKPEAKGAIN, attrs.value( "trackpeakgain" ).toString().toDouble() );
                     }
 
                     data.insert( Meta::Field::UNIQUEID, attrs.value( "uniqueid" ).toString() );
